+++ src/cpu/i960/i960dis.h
@@ -4,8 +4,8 @@
 typedef struct
 {
 	char		*buffer;	// output buffer
-	unsigned long	IP;
-	unsigned long	IPinc;
+	UINT32	IP;
+	UINT32	IPinc;
 } disassemble_t;
 
 char *i960_disassemble(disassemble_t *diss);
